Add files via upload
created a page with an HTML form, with JavaScript validation. This will be submitted as a single page with JavaScript, CSS, and HTML on the same page.
The assignment must have a page with a form on it that will be validated. The form must ask for required fields to be filled, including:
 Name (first and last). (required)
 Email address (required), which should be checked for proper format. For email validation, make sure:
         “@” symbol is present, and not in the first, second-last, or last positions of the string.
          “.” Character is present, and not in the first, second-last, or last positions of the string.
 Telephone number (required), which should be checked for proper format which is 12 characters, with hyphens: 204-555-1212).
 Postal Code (required), which should be checked for the proper number of characters (6 characters with no space).
 A full URL (not required but must be validated if present), which should be checked for proper format (starts with "http://")
 Age (not required), which must be validated if entered to make sure it is a number, not any other characters. Note: use a text field for better validation, not the number input type.

Note: "required" means the user is required to enter these fields. All the above must be present in your form, whether required or not.
This form should check to see that all the required fields have been filled. If they have not been filled or if any of the required fields are not in the proper format, the form should not submit and let the user know that the form is not filled incorrectly.
All form validation must work in JavaScript not HTML5. Be sure to add the novalidate attribute to your <form> tag.

2. JavaScript source code is well-formatted and easy to read.
3. Logical naming conventions and usages of variables, functions and IDs: All must be appropriately named and indicative of their purpose.
4. Use JavaScript comments wherever possible. Do not just comment because you must; explain what the code is doing.
5. In all cases, let the user know why the form is not validating, and how they can fix the problems with either alerts or innerHTML.